\name{beadarray-package}
\alias{beadarray-package}
\docType{package}
\title{
The beadarray package: a tool for low-level analysis of Illumina BeadArrays
}
\description{
The beadarray package: a tool for low-level analysis of Illumina BeadArrays
}


\author{
  Mark Dunning, Mike Smith, Jonathan Cairns, Matt Richie, Andy Lynch
}


\keyword{ package }
