\name{deprecatedFunctions}
\alias{createBeadSummaryData}
\alias{arrayNames}
\alias{getArrayData}
\alias{checkPlatform}
\alias{getAnnotation}
\alias{setAnnotation}

\title{Deprecated Functions}

\description{
Functions that have been renamed in the latest version of beadarray, but kept in for backwards compatibility.

}


\author{Mark Dunning}

\keyword{manip}
