\name{dim}
\docType{methods}
\alias{dim,beadLevelData-method}
\alias{dim,ExpressionSetIllumina-method}
\title{Retrieve the dimensions of an object}
\description{
	Retrieve the dimension of an object.
}
\usage{

\S4method{dim}{beadLevelData}(x)

\S4method{dim}{ExpressionSetIllumina}(x)

}
\arguments{
\item{x}{An object of class \code{\link[=beadLevelData-class]{beadLevelData}} or \code{\link[=ExpressionSetIllumina-class]{ExpressionSetIllumina}} }
}


\author{Mark Dunning}


\keyword{methods}
