\name{platformSigs}
\alias{platformSigs}
\docType{data}
\title{Annotation definitions}
\description{

A list of bead-level IDs for each Illumina expression array platform, as obtained by the lumi mapping packages. These are used to suggest the annotation of a possibly unknown bead-level dataset.

}
\usage{data(platformSigs)}
\keyword{datasets}
