\name{plotChipLayout}
\alias{plotChipLayout}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function to Plot the Layout of an Illumina BeadChip
}
\description{
Using the values obtained from a sentrix descriptor object generated by simpleXMLparse, a plot is generated showing samples and sections. 
}
\usage{
plotChipLayout(SD, subsC = NULL, sampC = NULL, sectC = NULL, desectC = "red", markC = "white", main = NULL, samplab = NULL, sectlab = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{SD}{
A sentrix descriptor object such as that generated by simpleXMLparse from an Illumina sdf file
}
  \item{subsC}{
The colour for the array substrate.  Will be extracted from the sentrix descriptor object if null.
}
  \item{sampC}{
The colour for the samples.  Will be extracted from the sentrix descriptor object if null.
}
  \item{sectC}{
The colour for the sections.  Will be extracted from the sentrix descriptor object if null.
}
  \item{desectC}{
The colour with which to outline the decode sections
}
  \item{markC}{
The colour for plotting decoding and analytical markers
}
  \item{main}{
A title to give the figure (e.g. chip name)
}
  \item{samplab}{
labels for the samples.  Will be extracted from the sentrix descriptor object if null.
}
  \item{sectlab}{
labels for the sections (if appropriate).  Will be extracted from the sentrix descriptor object if null.
}
}

\value{
returns a plot to the current device (recommend this is tall and thin)
}

\author{
Andy Lynch
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%\code{\link{simpleXMLparse}}
%}
\examples{
## SD<-simpleXMLparse(mysdf)
## plotChipLayout(SD)

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{hplot}

